requirements.txt
flask==2.3.3
langchain==0.1.0
langchain-openai==0.0.2
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0 